backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
openai
supabase
python-multipart
httpx
cachetools
redis>=5.0.0
slowapi>=0.1.9
pyyaml
requests
pyjwt[crypto]>=2.8.0
email-validator>=2.0.0
sentry-sdk[fastapi]>=2.0.0
tenacity>=8.2.0

